pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moppu"
version = "0.1.0"
description = "YouTube-context-driven trading agent with swappable LLM providers"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "khaneun" }]

dependencies = [
    # Core
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "typer>=0.12",
    "httpx>=0.27",
    "tenacity>=8.3",
    "structlog>=24.1",

    # YouTube ingestion
    "yt-dlp>=2024.8.6",
    "youtube-transcript-api>=0.6.2",
    "feedparser>=6.0.11",

    # Storage
    "sqlalchemy>=2.0",
    "alembic>=1.13",

    # Vector store / embeddings
    "chromadb>=0.5",
    "sentence-transformers>=3.0",

    # LLM providers
    "openai>=1.40",
    "anthropic>=0.34",
    "google-generativeai>=0.7",

    # Telegram
    "python-telegram-bot>=21.4",

    # Scheduling
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
moppu = "moppu.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/moppu"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
